+++ env/src/main/java/com/sap/cloud/security/config/OAuth2ServiceConfigurationBuilder.java
@@ -5,8 +5,6 @@
  */
 package com.sap.cloud.security.config;
 
-import com.sap.cloud.security.xsuaa.Assertions;
-
 import javax.annotation.Nonnull;
 import java.net.URI;
 import java.util.*;
@@ -35,7 +33,9 @@ public class OAuth2ServiceConfigurationBuilder {
 	 * @return this builder
 	 */
 	public static OAuth2ServiceConfigurationBuilder forService(@Nonnull Service service) {
-		Assertions.assertNotNull(service, "Service must not be null!");
+		if (service == null) {
+			throw new IllegalArgumentException("Service must not be null!");
+		}
 		OAuth2ServiceConfigurationBuilder instance = new OAuth2ServiceConfigurationBuilder();
 		instance.service = service;
 		return instance;